# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Anki Card Generator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Deck {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parentId?: number | null;
  cardCount: number;
  createdAt: string;
}

export interface CreateDeckBody {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parentId?: number | null;
}

export interface Card {
  id: number;
  deckId: number;
  front: string;
  back: string;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}

export interface UpdateCardBody {
  front?: string;
  back?: string;
  /** @nullable */
  tags?: string | null;
}

export interface GenerateCardsBody {
  text: string;
  deckName: string;
  cardCount?: number;
  /** @nullable */
  parentId?: number | null;
}

export interface GenerateCardsResponse {
  deck: Deck;
  cards: Card[];
  generatedCount: number;
}

export interface ExportDeckResponse {
  deckName: string;
  csv: string;
  cardCount: number;
}